Update converter class names
