+++ google-health-library/src/main/kotlin/org/radarbase/googlehealth/converter/GoogleHealthExerciseAvroConverter.kt
@@ -22,9 +22,10 @@ import org.radarbase.googlehealth.user.User
 import org.radarbase.googlehealth.util.googleHealthExerciseHeartRate
 import org.radarbase.googlehealth.util.googleHealthExercise
 import org.radarbase.googlehealth.util.googleHealthSource
+import java.io.IOException
 import java.time.Instant
 
-class ExerciseGoogleHealthAvroConverter(topic: String) : GoogleHealthAvroConverter(topic) {
+class GoogleHealthExerciseAvroConverter(topic: String) : GoogleHealthAvroConverter(topic) {
     override fun convertDataPoint(
         point: JsonNode,
         user: User,
@@ -76,12 +77,15 @@ class ExerciseGoogleHealthAvroConverter(topic: String) : GoogleHealthAvroConvert
             }
         }
 
-        val activityId = (point["name"] ?: point["dataPointName"])?.asText()
-            ?.substringAfterLast('/')?.toLongOrNull()
-            ?: run {
-                logger.warn("Dropping exercise data point with no usable log id for user={}", user.versionedId)
-                return emptyList()
-            }
+        val idSegment = (point["name"] ?: point["dataPointName"])?.asText()?.substringAfterLast('/')
+            ?: throw IOException(
+                "Exercise data point has no name or dataPointName to derive a log id from " +
+                    "for user=${user.versionedId}",
+            )
+        val activityId = idSegment.toLongOrNull()
+            ?: throw IOException(
+                "Exercise data point log id '$idSegment' is not numeric for user=${user.versionedId}",
+            )
 
         val record = googleHealthExercise {
             time = epochSeconds(start)
