+++ google-health-library/src/main/kotlin/org/radarbase/googlehealth/converter/GoogleHealthHeartRateAvroConverter.kt
@@ -19,21 +19,21 @@ package org.radarbase.googlehealth.converter
 import com.fasterxml.jackson.databind.JsonNode
 import org.apache.avro.specific.SpecificRecord
 import org.radarbase.googlehealth.user.User
-import org.radarbase.googlehealth.util.googleHealthOxygenSaturation
+import org.radarbase.googlehealth.util.googleHealthHeartRate
 
-class OxygenSaturationGoogleHealthAvroConverter(topic: String) :
-    GoogleHealthAvroConverter(topic) {
+class GoogleHealthHeartRateAvroConverter(topic: String) : GoogleHealthAvroConverter(topic) {
     override fun convertDataPoint(
         point: JsonNode,
         user: User,
     ): List<Pair<SpecificRecord, SpecificRecord>> {
-        val data = point["oxygenSaturation"] ?: return emptyList()
+        val data = point["heartRate"] ?: return emptyList()
         val time = parseSampleTime(data) ?: return emptyList()
-        val pct = data["percentage"]?.takeIf { it.isNumber }?.floatValue() ?: return emptyList()
-        val record = googleHealthOxygenSaturation {
+        val bpm = data["beatsPerMinute"]?.asInt() ?: return emptyList()
+        val record = googleHealthHeartRate {
             this.time = epochSeconds(time)
             timeReceived = nowEpochSeconds()
-            percentage = pct
+            timeInterval = 1
+            heartRate = bpm
         }
         return listOf(user.observationKey to record)
     }
