+++ google-health-library/src/main/kotlin/org/radarbase/googlehealth/converter/GoogleHealthSleepClassicAvroConverter.kt
@@ -19,32 +19,33 @@ package org.radarbase.googlehealth.converter
 import com.fasterxml.jackson.databind.JsonNode
 import org.apache.avro.specific.SpecificRecord
 import org.radarbase.googlehealth.user.User
-import org.radarbase.googlehealth.util.googleHealthSleepStage
-import org.radarcns.push.googlehealth.GoogleHealthSleepStageLevel
+import org.radarbase.googlehealth.util.googleHealthSleepClassic
+import org.radarcns.push.googlehealth.GoogleHealthSleepClassicLevel
 import java.time.Instant
 import java.time.LocalDateTime
 import java.time.ZoneOffset
 import java.time.format.DateTimeFormatter
 
-class SleepStageGoogleHealthAvroConverter(topic: String) : GoogleHealthAvroConverter(topic) {
+class GoogleHealthSleepClassicAvroConverter(topic: String) : GoogleHealthAvroConverter(topic) {
     override fun convertDataPoint(
         point: JsonNode,
         user: User,
     ): List<Pair<SpecificRecord, SpecificRecord>> {
         val data = point["sleep"] ?: return emptyList()
         val stages = data["stages"]?.takeIf { it.isArray } ?: return emptyList()
-        if (SleepSession.familyOf(data) != SleepSessionFamily.STAGES) return emptyList()
+        if (SleepSession.familyOf(data) != SleepSessionFamily.CLASSIC) return emptyList()
         val timeReceived = nowEpochSeconds()
         return stages.mapNotNull { stage ->
             val stageType = stage["type"]?.asText() ?: return@mapNotNull null
-            if (stageType !in STAGES_FAMILY) return@mapNotNull null
+            if (stageType !in CLASSIC_FAMILY) return@mapNotNull null
             val start = stage["startTime"]?.asText()
                 ?.let { runCatching { Instant.parse(it) }.getOrNull() } ?: return@mapNotNull null
             val end = stage["endTime"]?.asText()
                 ?.let { runCatching { Instant.parse(it) }.getOrNull() } ?: return@mapNotNull null
-
+            // Render in the stage's own UTC offset so dateTime is the device's local wall clock
+            // (like Fitbit), not UTC. Google derives its civil fields the same way (physical + offset).
             val startZone = ZoneOffset.ofTotalSeconds(parseUtcOffsetSeconds(stage["startUtcOffset"]?.asText()))
-            val record = googleHealthSleepStage {
+            val record = googleHealthSleepClassic {
                 dateTime = LOCAL_FMT.format(LocalDateTime.ofInstant(start, startZone))
                 this.timeReceived = timeReceived
                 duration = (end.epochSecond - start.epochSecond).toInt().coerceAtLeast(0)
@@ -54,16 +55,15 @@ class SleepStageGoogleHealthAvroConverter(topic: String) : GoogleHealthAvroConve
         }
     }
 
-    private fun mapLevel(text: String?): GoogleHealthSleepStageLevel = when (text) {
-        "DEEP" -> GoogleHealthSleepStageLevel.DEEP
-        "LIGHT" -> GoogleHealthSleepStageLevel.LIGHT
-        "REM" -> GoogleHealthSleepStageLevel.REM
-        "AWAKE" -> GoogleHealthSleepStageLevel.AWAKE
-        else -> GoogleHealthSleepStageLevel.UNKNOWN
+    private fun mapLevel(text: String?): GoogleHealthSleepClassicLevel = when (text) {
+        "ASLEEP" -> GoogleHealthSleepClassicLevel.ASLEEP
+        "RESTLESS" -> GoogleHealthSleepClassicLevel.RESTLESS
+        "AWAKE" -> GoogleHealthSleepClassicLevel.AWAKE
+        else -> GoogleHealthSleepClassicLevel.UNKNOWN
     }
 
     companion object {
         private val LOCAL_FMT: DateTimeFormatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME
-        private val STAGES_FAMILY = setOf("DEEP", "LIGHT", "REM", "AWAKE")
+        private val CLASSIC_FAMILY = setOf("ASLEEP", "RESTLESS", "AWAKE")
     }
 }
